+++ crates/cubecl-opt/src/analyses/mod.rs
@@ -1,6 +1,5 @@
 mod base;
-pub mod dominance_frontiers;
-pub mod dominators;
+pub mod dominance;
 pub mod liveness;
 pub mod post_order;
 pub mod writes;
